Added the starter files + extra customization
Added the html logo, Colored some of the texts and adding decoration on the left of 3 paragraphs.

Used html semantics layout.

Added the PDF download button at the bottom of the page, When clicks=ed it will open in a new tab as I used target="_blank".

Now going to customize the text layout.